Improve dynamic linking across layers
* wrapper script uses relative paths, so emit it at build
  time and then leave it alone in the postinstall script
* accept runtime dynamic library load path overrides
* handle venvs that are set up with a multi-step symlink chain

Implementation improvements for #38
