+++ src/venvstacks/_injected/postinstall.py
@@ -10,8 +10,6 @@ This post-installation script is automatically injected when packing environment
 
 import json
 import os
-import shlex
-import sys
 
 from compileall import compile_dir
 from os.path import abspath
@@ -162,79 +160,6 @@ def generate_sitecustomize(
     return sc_contents
 
 
-def generate_python_sh(
-    env_python_path: Path,
-    dynlib_paths: Sequence[Path],
-) -> tuple[str, Path] | None:
-    """Generate a Python wrapper script that sets the dynamic linking config appropriately."""
-    if not dynlib_paths:
-        # No need to wrap the link if the dynamic library loading doesn't need customisation
-        return None
-    if not env_python_path.is_symlink():
-        # Symlink has already been replaced, don't try to replace it again
-        return None
-    check_absolute_path(env_python_path)
-    check_absolute_paths(dynlib_paths)
-    # Use file relative paths so the wrapper script is portable (e.g. for local-export)
-    # We avoid `walk_up=True` here to maintain Python 3.11 compatibility,
-    # and to limit the relative paths to remaining within peer environments
-    env_bin_dir_path = env_python_path.parent
-    env_bin_dir_name = env_bin_dir_path.name
-    env_path = env_bin_dir_path.parent
-    env_name = env_path.name
-    parent_path = env_path.parent
-
-    def _sh_path(p: Path) -> str:
-        parent_relative_path = p.relative_to(parent_path)
-        if parent_relative_path.parts[0] != env_name:
-            # Path is in a peer environment
-            path_parts = ("..", "..", *parent_relative_path.parts)
-        elif parent_relative_path.parts[1] != env_bin_dir_name:
-            # Path is in the same environment as the wrapper script
-            path_parts = ("..", *parent_relative_path.parts[1:])
-        else:
-            # Path is in the same directory as the wrapper script
-            path_parts = parent_relative_path.parts[2:]
-        return shlex.quote(str(Path(*path_parts)))
-
-    # Wrapper needs "exec -a" support, so specify a shell that provides it
-    # (most notably, "dash", which provide "/bin/sh" on Ubuntu omits it)
-    # Also set the relevant environment variable for dynamic loading config
-    if sys.platform == "darwin":
-        shell = "/bin/zsh"
-        dynlib_var = "DYLD_LIBRARY_PATH"
-    else:
-        shell = "/usr/bin/bash"
-        dynlib_var = "LD_LIBRARY_PATH"
-
-    dynlib_content: list[str] = [
-        # Based on the PATH manipulation suggestion in https://unix.stackexchange.com/a/124447
-        # Path list is reversed on iteration because the helper function *pre*pends each entry
-        f'add_dynlib_dir() {{ case ":${{{dynlib_var}:=$1}}:" in *:"$1":*) ;; *) {dynlib_var}="$1:${dynlib_var}" ;; esac; }}',
-        *(
-            f'add_dynlib_dir "$script_dir/{_sh_path(p)}"'
-            for p in reversed(dynlib_paths)
-        ),
-        f"export {dynlib_var}",
-    ]
-    symlink_path = env_bin_dir_path / f"{env_python_path.name}_"
-    wrapper_script_lines = [
-        f"#!{shell}",
-        "# Allow extension modules to find shared libraries published by lower layers",
-        "set -eu",
-        "# Note: `readlink -f` (long available in GNU coreutils) is available on macOS 12.3 and later",
-        'script_dir="$(cd "$(dirname "$(readlink -f "$0")")" 1> /dev/null 2>&1 && pwd)"',
-        *dynlib_content,
-        f'script_path="$script_dir/{_sh_path(env_python_path)}"',
-        f'symlink_path="$script_dir/{_sh_path(symlink_path)}"',
-        'test -f "$script_path" || echo 1>&2 "Invalid wrapper script path: $script_path"',
-        'test -L "$symlink_path" || echo 1>&2 "Invalid base Python symlink: $symlink_path"',
-        'exec -a "$script_path" "$symlink_path" "$@"',
-    ]
-    wrapper_script_contents = "\n".join(wrapper_script_lines)
-    return wrapper_script_contents, symlink_path
-
-
 def _run_postinstall(layer_path: Path) -> None:
     """Run the required post-installation steps in a deployed environment."""
     # Read the layer config file
@@ -244,31 +169,18 @@ def _run_postinstall(layer_path: Path) -> None:
     env_python_path = config["python_path"]
     if base_python_path != env_python_path:
         # Generate `pyvenv.cfg` for layered environments
+        # (path to base python must be emitted as an absolute path)
         venv_config = generate_pyvenv_cfg(base_python_path, config["py_version"])
         venv_config_path = layer_path / "pyvenv.cfg"
         venv_config_path.write_text(venv_config, encoding="utf-8")
 
         # Generate `sitecustomize.py` for layered environments
+        # (avoids having to resolve relative paths on every startup)
         dynlib_paths = config["dynlib_paths"]
         sc_contents = generate_sitecustomize(config["pylib_paths"], dynlib_paths)
         if sc_contents is not None:
             sc_path = config["site_path"] / "sitecustomize.py"
             sc_path.write_text(sc_contents, encoding="utf-8")
-
-        # Potentially generate Python wrapper script for non-Windows envs
-        # We still run via the symlink so the path to pyvenv.cfg is valid
-        # regardless of whether or not "exec -a" is taken into account
-        if dynlib_paths and not hasattr(os, "add_dll_directory"):
-            wrapper_info = generate_python_sh(
-                env_python_path,
-                dynlib_paths,
-            )
-            if wrapper_info is not None:
-                sh_contents, symlink_path = wrapper_info
-                # Replace the symlink with a wrapper script
-                env_python_path.rename(symlink_path)
-                env_python_path.write_text(sh_contents, encoding="utf-8")
-                os.chmod(env_python_path, 0x755)
 
     # Precompile Python library modules
     pylib_path = (
